refactor: Rename `…lugins/aws/package/compile/events/cloudWatchEvent.js`
