+++ lib/plugins/aws/package/compile/events/cloud-watch-event.js
