+++ lib/plugins/index.js
@@ -46,7 +46,7 @@ module.exports = [
   require('./aws/package/compile/events/alexa-smart-home.js'),
   require('./aws/package/compile/events/iot.js'),
   require('./aws/package/compile/events/iotFleetProvisioning.js'),
-  require('./aws/package/compile/events/cloudWatchEvent.js'),
+  require('./aws/package/compile/events/cloud-watch-event.js'),
   require('./aws/package/compile/events/cloudWatchLog.js'),
   require('./aws/package/compile/events/cognitoUserPool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
